Finished custom hook creation. Added spme css styling
