+++ src/components/index.js
@@ -1,0 +1,3 @@
+export * from "./AddCategory";
+export * from "./GifGrid";
+export * from "./GifItem";
