+++ src/helpers/getGifs.js
@@ -1,10 +1,12 @@
 const getGifs = async(category) => {
-    const url = `https://api.giphy.com/v1/gifs/search?api_key=FWCszwwyG9YJM5R3VACQkGLUQBl2M7Qe&q=${category}&limit=10`;
+    const url = `https://api.giphy.com/v1/gifs/search?api_key=FWCszwwyG9YJM5R3VACQkGLUQBl2M7Qe&q=${category}&limit=12`;
     const response = await fetch(url, {
         method: "GET",
         mode: "cors"
     });
     const {data} = await response.json();
+    console.log(data);
+
     const gifs = data.map((gif) => {
         return {
             id: gif.id,
