+++ src/styles.css
@@ -1,10 +1,11 @@
 * {
     font-family: Helvetica, Arial, sans-serif;
-    background-color: rgb(234, 234, 234);
   }
   
   body {
     padding: 60px;
+    background-color: rgb(234, 234, 234);
+
   }
   form {
     display: flex;
@@ -50,44 +51,76 @@
     flex-direction: row;
     flex-wrap: wrap;
     justify-content: center;
+    animation: slide;
+    animation-duration: 1s;
+    transition: ease-in-out
+  }
+
+  @keyframes slide {
+    from{
+      margin-top: 100%;
+      display: flex;
+      flex-direction: row;
+      flex-wrap:nowrap;
+    }
+
+    to{
+      margin-top: 0%;
+      display: flex;
+      flex-direction: row;
+      flex-wrap:wrap;
+    }
+    
   }
   
   .card {
-    align-content: center;
-    align-items: center;
     background-color: white;
     border-radius: 10px;
     box-shadow: 10px 10px 0px rgba(0, 0, 0, 0.251);
     display: flex;
     flex-direction: column;
-    width: 480px;
-    height: 350px;  
-    justify-content: center;
-    align-items: flex;
+ 
+
+    width: 380px;
+    height: 250px;  
+
     margin-bottom: 30px;
     margin-right: 40px;
     overflow: hidden;
+
+    position: relative;
   }
   
-  .card p {
-    background-color: white;
+  .card span {
     display: flex;
+    flex-direction: column;
     align-items: center;
+    justify-content: flex-end;
+   
+    width: inherit;
+    height: inherit;
+    color:transparent;
     flex: 1;
     font-size: 1.5rem;
-    margin-top: 10px;
-    margin-bottom: 10px;
-    padding: 5px;
     text-align: center;
+    position: absolute;
+    margin: 0px;
+
+    text-transform: capitalize;
   }
   
   .card img {
     width: 100%;
-    height: 75%;
+    height: 100%;
   }
 
   .card:hover{
     transform: scale(1.05);
     transition: ease-in-out 0.15s;
     box-shadow: 10px 10px 0px rgba(233, 69, 96, 0.251);
+  }
+
+  .card span:hover{
+    color: white;
+    text-shadow: 2px 2px #E94560;
   }
